+++ pac4j-jwt/src/main/java/org/pac4j/jwt/config/signature/SecretSignatureConfiguration.java
@@ -2,13 +2,11 @@ package org.pac4j.jwt.config.signature;
 
 import com.nimbusds.jose.JOSEException;
 import com.nimbusds.jose.JWSAlgorithm;
-import com.nimbusds.jose.JWSHeader;
 import com.nimbusds.jose.JWSSigner;
 import com.nimbusds.jose.JWSVerifier;
 import com.nimbusds.jose.crypto.MACSigner;
 import com.nimbusds.jose.crypto.MACVerifier;
 import com.nimbusds.jose.util.Base64;
-import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
 import lombok.val;
 import org.pac4j.core.exception.TechnicalException;
@@ -94,14 +92,9 @@ public class SecretSignatureConfiguration extends AbstractSignatureConfiguration
 
     /** {@inheritDoc} */
     @Override
-    public SignedJWT sign(final JWTClaimsSet claims) {
-        init();
-
+    protected JWSSigner buildSigner() {
         try {
-            JWSSigner signer = new MACSigner(this.secret);
-            val signedJWT = new SignedJWT(new JWSHeader(algorithm), claims);
-            signedJWT.sign(signer);
-            return signedJWT;
+            return new MACSigner(this.secret);
         } catch (final JOSEException e) {
             throw new TechnicalException(e);
         }